Add files via upload
Employee turn-over otherwise known as "employee churn" is a costly
problem for companies. The cost of replacing employees for most employers remains
significant. This is due to the amount of time spent on the interviews and find a replacement,
sign-on bonuses, and the loss of productivity for several months while the new employee gets
accustomed to the new role.
In this project “Human Resource analytics and Employee Churn Prediction” we plan to start
off with Descriptive Analytics and Exploratory data analysis and then apply predictive
modelling in an attempt to solve our problem statements.